+++ CHANGELOG.md
@@ -8,10 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [11.12.0] - 2023-10-XX
 
 * Add PHP 8.3 support
-* Add support for `environment_scope` in `Project::removeVariable`
+* Add `Projects::updateProtectedBranch` and `Projects::updateApprovalsConfiguration`
+* Add support for `environment_scope` in `Projects::removeVariable`
+* Add support for `filter` in `Projects::variable`
 * Add support for `author` in `Repositories::commits`
+* Add support for additional parameters in `Projects::labels` and `Groups::labels`
 
-## [11.11.1] - 2023-10-XX
+## [11.11.1] - 2023-10-08
 
 * Fixed double encoding of job name in artifacts download